Last commits

+++ mosaicode/GUI/diagram.py
@@ -3,8 +3,6 @@
 This module contains the Diagram class.
 """
 import gi
-from copy import deepcopy
-from copy import copy
 gi.require_version('Gtk', '3.0')
 gi.require_version('GooCanvas', '2.0')
 from gi.repository import Gtk
@@ -90,9 +88,10 @@ class Diagram(GooCanvas.Canvas, DiagramModel):
 
     # ----------------------------------------------------------------------
     def __on_key_press(self, widget, event=None):
-        grid = System.properties.grid
+        grid = System.get_preferences().grid
         modifier_mask = Gtk.accelerator_get_default_mod_mask()
         event.state = event.state & modifier_mask
+
         if event.state == Gdk.ModifierType.CONTROL_MASK:
             if event.keyval == Gdk.KEY_Up:
                 self.move_selected(0, -grid*5)
@@ -107,7 +106,7 @@ class Diagram(GooCanvas.Canvas, DiagramModel):
                 self.move_selected(grid*5, 0)
                 return True
 
-        if event.keyval == Gdk.KEY_Delete:
+        if event.keyval == Gdk.KEY_Delete and self.focus:
             self.delete()
             return True
 
@@ -132,12 +131,15 @@ class Diagram(GooCanvas.Canvas, DiagramModel):
     def __on_button_press(self, widget, event=None):
         Gtk.Widget.grab_focus(self)
         if event.button == 1:
-            self.main_window.main_control.show_diagram_property(self)
+            self.main_window.property_box.set_diagram(self)
             self.last_clicked_point = (event.x, event.y)
             self.deselect_all()
             self.__abort_connection()
             self.update_flows()
             self.__start_select()
+            return False
+        elif event.button == 3:
+            self.main_window.diagram_menu.show(self, event)
             return False
         return False
 
@@ -296,7 +298,7 @@ class Diagram(GooCanvas.Canvas, DiagramModel):
                         data="M 0 " + str(i) + " L "+ str(width) +" "+ str(i) + "",
                         line_width=0.8
                         )
-                i = i + System.properties.grid
+                i = i + System.get_preferences().grid
             i = 0
             while i < width:
                 GooCanvas.CanvasPath(
@@ -305,12 +307,7 @@ class Diagram(GooCanvas.Canvas, DiagramModel):
                         data="M " + str(i) + " 0 L "+ str(i) + " "+ str(height) +"",
                         line_width=0.8
                         )
-                i = i + System.properties.grid
-
-    # ----------------------------------------------------------------------
-    def set_show_grid(self, bool):
-        if bool is not None:
-            self.show_grid = bool
+                i = i + System.get_preferences().grid
 
     # ----------------------------------------------------------------------
     def update_flows(self):
@@ -324,18 +321,6 @@ class Diagram(GooCanvas.Canvas, DiagramModel):
             conn.update_flow()
         for comment in self.comments:
             comment.update_flow()
-
-    # ----------------------------------------------------------------------
-    def set_file_name(self, file_name):
-        """
-        This method set name of diagram file.
-
-            Parameters:
-                * **file_name** (:class:`str<str>`)
-
-        """
-        self.file_name = file_name
-        self.main_window.work_area.rename_diagram(self)
 
     # ----------------------------------------------------------------------
     def change_zoom(self, value):
@@ -365,7 +350,7 @@ class Diagram(GooCanvas.Canvas, DiagramModel):
             Parameters:
                 * **block**(:class: `Block<mosaicode.GUI.block>`)
         """
-        self.main_window.main_control.show_comment_property(comment)
+        self.main_window.property_box.set_comment(comment)
 
     # ----------------------------------------------------------------------
     def show_block_property(self, block):
@@ -375,7 +360,7 @@ class Diagram(GooCanvas.Canvas, DiagramModel):
             Parameters:
                 * **block**(:class: `Block<mosaicode.GUI.block>`)
         """
-        self.main_window.main_control.show_block_property(block)
+        self.main_window.property_box.set_block(block)
 
     # ----------------------------------------------------------------------
     def resize(self, data):
@@ -435,6 +420,14 @@ class Diagram(GooCanvas.Canvas, DiagramModel):
             comment.move(x, y)
         self.update_flows()
 
+    # ----------------------------------------------------------------------
+    def collapse(self, state):
+        for key in self.blocks:
+            if not self.blocks[key].is_selected:
+                continue
+            self.blocks[key].is_collapsed = state
+        self.update_flows()
+
     # ---------------------------------------------------------------------
     def check_limit(self, x, y, block_pos_x, block_pos_y):
         min_x = 0
@@ -457,94 +450,6 @@ class Diagram(GooCanvas.Canvas, DiagramModel):
 
         return x, y
 
-    # ---------------------------------------------------------------------
-    def paste(self):
-        """
-        This method paste a block.
-        """
-        replace = {}
-        self.deselect_all()
-        # interact into blocks, add blocks and change their id
-        clipboard = self.main_window.main_control.get_clipboard()
-        for widget in clipboard:
-            if not isinstance(widget, Block):
-                continue
-            block = BlockModel(widget)
-            block.x += 20
-            block.y += 20
-            block.id = -1
-            if not self.main_window.main_control.add_block(block):
-                return
-            replace[widget.id] = block
-            block.is_selected = True
-        # interact into connections changing block ids
-        for widget in clipboard:
-            if not isinstance(widget, Connector):
-                continue
-            # if a connector is copied without blocks
-            if widget.output.id not in replace or widget.input.id \
-                    not in replace:
-                continue
-            output = replace[widget.output.id]
-            output_port = widget.output_port
-            input = replace[widget.input.id]
-            input_port = widget.input_port
-            self.start_connection(output, output_port)
-            self.curr_connector.is_selected = True
-            self.end_connection(input, input_port)
-        self.update_flows()
-
-    # ---------------------------------------------------------------------
-    def copy(self):
-        """
-        This method copy a block.
-        """
-        self.main_window.main_control.reset_clipboard()
-        for key in self.blocks:
-            if not self.blocks[key].is_selected:
-                continue
-            self.main_window.main_control.get_clipboard().append(self.blocks[key])
-        for conn in self.connectors:
-            if not conn.is_selected:
-                continue
-            self.main_window.main_control.get_clipboard().append(conn)
-        for comment in self.comments:
-            if not comment.is_selected:
-                continue
-            self.main_window.main_control.get_clipboard().append(comment)
-
-    # ---------------------------------------------------------------------
-    def cut(self):
-        """
-        This method delete a block.
-        """
-        self.do(_("Cut"))
-        self.self.copy()
-        self.delete()
-
-    # ---------------------------------------------------------------------
-    def delete(self):
-        """
-        This method delete a block or connection.
-        """
-        self.do("Delete")
-        for key in self.blocks.copy():
-            if not self.blocks[key].is_selected:
-                continue
-            del self.blocks[key]
-        for con in self.connectors:
-            if not con.is_selected:
-                continue
-            if con not in self.connectors:
-                continue
-            self.connectors.remove(con)
-        for comment in self.comments:
-            if not comment.is_selected:
-                continue
-            self.comments.remove(comment)
-
-        self.deselect_all()
-        self.redraw()
 
     # ---------------------------------------------------------------------
     def set_modified(self, state):
@@ -595,8 +500,8 @@ class Diagram(GooCanvas.Canvas, DiagramModel):
         for comment in self.comments:
             if not isinstance(comment, Comment):
                 comm = Comment(self)
-                comm.set_text(comment.text)
                 comm.move(comment.x, comment.y)
+                comm.update_flow()
                 self.comments[i] = comm
             i = i + 1
 
@@ -625,88 +530,12 @@ class Diagram(GooCanvas.Canvas, DiagramModel):
         # Redraw Comments
         for comment in self.comments:
             self.get_root_item().add_child(comment, -1)
+            comment.adjust_position()
 
-        self.update_flows()
-
-    # ---------------------------------------------------------------------
-    def do(self, new_msg):
-        """
-        This method do something
-            Parameters:
-                * **new_msg** (:class:`str<str>`)
-        """
-        self.set_modified(True)
-        action = (copy(self.blocks),    #0
-                  copy(self.connectors),#1
-                  copy(self.comments),  #2
-                  new_msg)              #3
-        self.undo_stack.append(action)
-
-    # ---------------------------------------------------------------------
-    def undo(self):
-        """
-        This method undo a modification.
-        """
-        if len(self.undo_stack) < 1:
-            return
-        self.set_modified(True)
-        action = self.undo_stack.pop()
-        self.blocks = action[0]
-        self.connectors = action[1]
-        self.comments = action[2]
-        msg = action[3]
-        self.redraw()
-        self.redo_stack.append(action)
-        if len(self.undo_stack) == 0:
-            self.set_modified(False)
-
-    # ---------------------------------------------------------------------
-    def redo(self):
-        """
-        This method redo a modification.
-        """
-        if len(self.redo_stack) < 1:
-            return
-        self.set_modified(True)
-        action = self.redo_stack.pop()
-        self.blocks = action[0]
-        self.connectors = action[1]
-        self.comments = action[2]
-        msg = action[3]
-        self.redraw()
-        self.undo_stack.append(action)
-
-    # ----------------------------------------------------------------------
-    def align(self, alignment):
-        top = self.main_window.get_size()[1]
-        bottom = 0
-        left = self.main_window.get_size()[0]
-        right = 0
-
-        for key in self.blocks:
-            if not self.blocks[key].is_selected:
-                continue
-            x, y = self.blocks[key].get_position()
-            if top > y: top = y
-            if bottom < y: bottom = y
-            if left > x: left = x
-            if right < x: right = x
-
-        for key in self.blocks:
-            if not self.blocks[key].is_selected:
-                continue
-            x, y = self.blocks[key].get_position()
-            if alignment == "BOTTOM":
-                self.blocks[key].move(0, bottom -y)
-            if alignment == "TOP":
-                self.blocks[key].move(0, top -y)
-            if alignment == "LEFT":
-                self.blocks[key].move(left -x, 0)
-            if alignment == "RIGHT":
-                self.blocks[key].move(right -x, 0)
         self.update_flows()
 
     # ----------------------------------------------------------------------
     def show_block_menu(self, block, event):
-        self.main_window.block_menu.show_block_menu(block, event)
+        self.main_window.block_menu.show(block, event)
+
 # ----------------------------------------------------------------------
